# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Swagger-generated docs for My API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Tạo đơn hàng mới",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "products",
                                    "payment_method"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "description": "ID của profile người dùng",
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "description": "ID của phương thức thanh toán",
                                        "type": "integer"
                                    },
                                    "products": {
                                        "description": "Danh sách sản phẩm trong đơn hàng",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_variant_id"
                                            ],
                                            "properties": {
                                                "product_variant_id": {
                                                    "description": "ID của biến thể sản phẩm",
                                                    "type": "integer"
                                                },
                                                "serial": {
                                                    "description": "Số serial của sản phẩm (sẽ được tự động tạo nếu không có)",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo đơn hàng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID của đơn hàng",
                                                    "type": "integer"
                                                },
                                                "profile_id": {
                                                    "description": "ID của profile người dùng",
                                                    "type": "integer"
                                                },
                                                "status": {
                                                    "description": "Trạng thái đơn hàng",
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "payment_method": {
                                                    "description": "ID của phương thức thanh toán",
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "description": "Thời gian tạo đơn hàng",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Thời gian cập nhật đơn hàng gần nhất",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "profile_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "payment_method": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "products.*.product_variant_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/check-status": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Kiểm tra tình trạng thanh toán của đơn hàng",
                "operationId": "checkOrderStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "description": "ID của đơn hàng cần kiểm tra",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header sử dụng cơ chế Bearer.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}